requirements.txt
# Indian Stock Market Analytics - Python Dependencies

# Core Data & Analysis
pandas>=1.5.0
numpy>=1.24.0
yfinance>=0.2.28
nsepy>=0.8

# Machine Learning
scikit-learn>=1.3.0
tensorflow>=2.13.0
keras>=2.13.0

# FinBERT and Transformers for Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
accelerate>=0.20.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Database & Caching
redis>=4.6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Data Visualization (Optional - for backend charts)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.16.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=11.0

# Async Support
aiohttp>=3.8.0
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Deployment
gunicorn>=21.2.0
uvicorn>=0.23.0

# Monitoring & Logging
python-json-logger>=2.0.7
sentry-sdk>=1.30.0

# Task Queue (Optional)
celery>=5.3.0
